requirements.txt
PyQt5==5.15.9
python-telegram-bot==20.5